requirements.txt
#duckdb
pillow
mutagen
pyacoustid
musicbrainzngs
requests